refactor(frontend): extract optimization form and complete PR4 split
Move the new-run form into optimization-form.tsx, slim the screen shell,
and centralize isRunnableDataset in optimization-model for thermo follow-up.
